// generated/src/app/api-models/inst-claim-detail-remarks.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class InstClaimDetailRemarks {
  seqClaimId: number;
  lineNumber: number;
  subLineCode: string;
  remarkCode: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}